Refactors continued
* Fix linting warning in ComponentIsOnFire
* Move handling of selected file to a utility function
